+++ src/datavisualization/engine/q3dcamera_p.h
@@ -44,14 +44,14 @@ public:
 
     void sync(Q3DCamera &other);
 
-    void setXRotation(qreal rotation);
-    void setYRotation(qreal rotation);
-    void setMinXRotation(qreal rotation);
-    void setMinYRotation(qreal rotation);
-    void setMaxXRotation(qreal rotation);
-    void setMaxYRotation(qreal rotation);
+    void setXRotation(float rotation);
+    void setYRotation(float rotation);
+    void setMinXRotation(float rotation);
+    void setMinYRotation(float rotation);
+    void setMaxXRotation(float rotation);
+    void setMaxYRotation(float rotation);
 
-    void updateViewMatrix(qreal zoomAdjustment);
+    void updateViewMatrix(float zoomAdjustment);
 
 public:
     Q3DCamera *q_ptr;
